conditions: fix parsing expression statements with a cast
When smatch sees an expression statement like:

foo = ({frob(); frob(); frob(); 1;});

Then it tries to parse the frob(); frob(); frob(); and create a fake
"foo = 1;" assignment.  The problem arises when there is a cast in front
of the expression statement:

	foo = (void *)({frob(); frob(); frob(); 1;});

I wrote the original code this way because originally strip_expr() would
have stripped away the cast and the ().  Also, although I didn't care about
it at the time, it's not right to ignore casts.

To fix it, I've created a new strip_expr_cast() function which separates
out the casts and returns them.  (There would be a string of casts like
"(void *)(unsigned long)").  Then we add the casts back as the final
step:
	foo = (void *)(unsigned long)1;

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>