flow: bail out earlier on parsing giant arrays
The drivers/net/wireless/realtek/rtw89/rtw8851b_table.c file has a bunch
of arrays which are giant.  The first one is 976 elements.  When you were
doing an --info run then it took Smatch over an hour to parse that file.

And actually the information we're saving is kind of garbage.  Instead of
saying 0-s32max it ends up being some list of numbers with a ... on the
end. "0,4,7,1999, 2000, ...".

This function does a weird thing where it treats globals differently from
local variables.  I feel like I probably did that because globals are
often huge and parsing them is slow...  But really I think that ideally
we would use __split_expr().

Anyway, what I've done here is that I've said if the array is more than 256
elements then do:

	my_array[unknown()] = unknown();

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>