implied: fix leaf checking
Hm...  I thought I had committed this.  The interesting part of this is
in get_tf_stacks_from_pool().  What happens is that when we're parsing
return statements then we take the string "ret = 1" or "ret = 0-u64max"
and then we look at all the leaf nodes where that is true.

However, actually it wasn't checking leaf nodes, it was checking merged
nodes as well.  And when we merge 1 and 0-u64max then the merged range
is still 0-u64max.  So we end up saying there basically are no implication
on that path.  (The 0-u64max value might not be the interesting bit, it's
just a marker for the path and there might be interesting implications on
that path).

The is_merged() -> is_leaf() change in sm_in_keep_leafs() is just for
correctness and might not fix a real life bug at all.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>