struct_assignment: use the ->zeroed flag from allocations
The struct assignment tracks when we assign a zeroed buffer to a struct.
Basically the current code said that kcalloc() returns zero and in the
kernel if it has "zalloc" name then it's zero.  This works when it works,
but now we have better data in smatch_allocations.c so we can use that
instead.

I had to add support for early allocation hooks to do this.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>