comparison: add support for COMPARE_LIMIT
I'm committing this code long after I actually made this change so I had
to look up what the logic is.  The comments are kind of crap...  The
difference between PARAM_COMPARE and COMPARE_LIMIT is PARAM_COMPARE only
looks at unmodified parameters where COMPARE_LIMIT looks at the values at
the end of the function as in "*offset <= length".

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>