implied: try again in get_tf_stacks_from_pool()
This reverts the code back almost to what it was a month ago or whatever...

The problem that you can have a merged state which looks the same as a
leaf state but we only wanted to deal with leaf states.  So add a check
for that.  Simple enough.

But then it only makes sense to check the ->possible list for merged
states.  Skipping that some how broke the parsing of the contains_operator()
function.

So this change basically takes it back to the old code except for the
new is_leaf() check.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>