kernel.return_fixes: add mipi_dsi_device_transfer(), timer_delete() and get_device()
The interesting function here is get_device().  It returns the parameter
that was passed in but it does some complicated pointer math to get there.

The problem with mipi_dsi_device_transfer() is the typical thing where
it returns an unknown long, but actually it can't return negatives except
error codes or positives above INT_MAX.

For some reason Smatch was parsing timer_delete() as returning zero.  Was
this a fundamental problem or because of the bugs in smatch_implied.c???

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>