mixing_irq_and_irqsave: warn about mixing _irq() and _irqsave()
It seldom makes sense to mix spinlock_irq() and spinlock_irqsave().  The
one is for when we know IRQs are not disabled and the other is for when
they are.  It can't be both.

These warning are mostly not bugs, but just places which should be changed
to _irq().  Except changing to _irgsave() is potentially safer.  The point
is the either works.  It's just sloppy and wrong to mix them.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>