leaf_fn: be stricter with PARAM_USED
Before the PARAM_USED was kind of mean "This is probably not used".  Smatch
would filter out variables which were zero.  (There are a lot of known
zero variables because of kzalloc()).

But it turned out we were passing ridiculous amounts of data to functions
that only did "return foo + 1; or whatever.  So lets mark "leaf" functions
where we're sure that no recursion, function pointers, or whatever is
involved as leaf functions.  For these functions we'll only pass exactly
what is needed.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>