db: don't load silly amounts of caller info
The issue here was is_inode_flag_set().  We say functions which are called
over 200 times as too common.  Don't save anything about them.  But
is_inode_flag_set() was called 166 times.

Apparently, there is no limit to the amount of caller info per call which
gets saved in the database.  This is hard to limit in a sensible ways as
well.

So is_inode_flag_set() was loading 28k entries of caller_info.  Which is
too much.  Cap it at 5000.  (This number is picked by guessing what an
appropriate number is).

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>