kernel: ignore "->bio->bi_private"
Nothing good can come from tracking ->bi_private data.

This came from looking at iomap_readpage_iter() returns states.  The
unfortunate thing is that most of this is from PARAM_USED.  It's possible
that it is all used, or that there is a bug where too much stuff is marked
as used.  Or it could be that there is recursion and that an old bug was
marking it as used but recursion means we don't get rid of it...

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>