preempt: introduce a new preempt_table to edit out caller_info() data
How this works is you give a function and a caller and that specific
function call is now not marked as calling with PREMPT_DISABLED.
Of course, if a function calls the same function twice both calls are now
marked as not-preempt.  It's not perfect but it's fine grained enough.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>