checking_for_null_instead_of_err_ptr: silence IS_ERR_OR_NULL() warnings
This check is only enabled when --spammy is turned on.  It complains when
a potential error pointer is checked against NULL.  Of course, pointers
can sometimes be either NULL or error pointers.  Which is why the
IS_ERR_OR_NULL() function exists.

I don't see this warning on my system, but it's simple enough to silence
it.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>