preempt_info: hard code some functions
class_task_rq_lock_destructor() does that if statement thing which is
not really an if statement in real life.

affine_move_task() does some slightly tricky stuff with regards to
preemption.  It bumps the counter, then it calls task_rq_unlock() which
decrements the counter, then it decrements the counter.  This check is
deliberately written so complicated things like that have to be handled
manually.  It gets pretty complicated to track multiple decrements and
increments automatically and I've never found a way to do it correctly.
This way works, it's still a headache to manually add functions like
affine_move_task() but it's less of a headache.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>