flow: reset position after calling cleanup functions
When there is a cleanup function then what happens is that Smatch
generates a fake function call.  The issue is that the fake call messes
up the current position.  So if you're planning to print an warning
message for that return, it instead get printed on the line where the
cleanup function was declared.

Save the current position, and reset it back to where it was after calling
the cleanup function.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>