flow: introduce stmt_get_parent_expr()
We want to be able to track statements back to their parents, but this
wasn't tracking expressions which hold statements.

Notice that the if (expr->statement && !expr->statement) { is dead code.
I guess the intent was to set the parent statement to whatever the most
recent statement was.  But I guess that caused a crash, and I thought I
solved the problem, but actually I solved it my making it into dead code.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>