db: export get_fake_return_variable()
When you return a function call:

	return frob();

Smatch translates that to:

	return __fake_return_23424234 = frob();

It's a bit tricky because the fake variable is hidden off to the side.
You need to call get_fake_return_variable() to see what states are
stored there.  Export the fake variable so other code can use it, not
just smatch_db.c.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>