db: change how db_incomplete() is tracked
The problem that db_incomplete() solves is imagine we do a select state
and there were no return values.  That means that we can't rely on
information from the database because clearly it isn't working well.

However, what happens is when I'm debugging I add an
#include "check_debug.h" and none of those functions are in the database
so now the database is "incomplete".  (Wait?  Why weren't these parsed
inline?)

So instead we set a flag for each function to say if it has parsed a call
with no data in the DB.

Anyway the problem is when you parse an inline function and that function
make calls  which don't have an entry in the DB the the DB can't be
trusted.  So it needs to be that if we or any inline functions parses a
call with no data in the DB then it's incomplete.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>