param_set: make kernel.ignore_uninitialized_param as PARAM_SET
I know there is a specific reason for this commit.  I can think of three
options in order of most to least likely:  It silences false positives.
Results from the database look more sensible.  It saves space in the DB.

I'm honestly not sure why I wrote this but it looks reasonable.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>