points_to_user_data: clear the state for "*p = 0;"
Smatch user data tracking uses two things.  We first check
smatch_kernel_user_data.c which tracks values and then
smatch_points_to_user_data.c which just says that the pointer has user
data.

So if there is a clamp on the values it works because that's tracked by
smatch_kernel_user_data.c.

The problem is that you could have "copy_from_user(p, src, sizeof(*p))"
it marks p as a user pointer.  (Writing this commit message, I think
actually it should only do that if p is an array not pointer to a single
int).  Anyway, then we set *p = 0.  Now user_data has no value and no
state so we default to saying that it's a user pointer and it means that
the user has total control.  When really the user has no control.

Clear the user pointer in this case.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>