locking: start to do check_locking re-write
I've been wanting to do a locking re-write for years.  When I wrote
check_locking.c I was quite proud of the locking table and that was
kind of the start of the param/key API.  However, what I've realized in
the time since is that it's better to break apart the module that records
state transitions into return_states and the module the records the
current state into caller_info.  Then break apart the actual checks into
separate modules and provide hooks for those.

The check_locking.c code was a mess because it was hard to tell what the
start state was coming into the function, but in the new code tracking the
callers is moved into a different module so that's much simpler.
(Everything starts as no state).

I had initially hoped that I could do something like what the sleeping
in preempt count does, but it turns out that sleeping in preempt is biased
towards saying we're not in preempt.  And also preempt is much simpler
because it's just one on/off instead of multiple different types of locks.
The locking code has to be much more accurate.

For example in check_locking.c, mutex_destroy() was treated as a restore
because we didn't have a double unlock warning.  This silenced warnings
about forgetting to unlock on error paths but it breaks the "double
unlock" warning.  In the new code it's tracked as a destroy.

There are a lot of hooks that I want to put into check_locking.c which are
difficult because it's written to a specific check and it's all entwined
together.  For example, it's sometimes difficult to do the param mapping
so maybe it would be better to say we locked "(struct foo)->bar" and
we unlocked "(struct foo)->bar", the names might be different but they're
probably the same.  That could be done in different modules using hooks.
Or a module which tracks when IRQs are disable or not.  The check for
double locking is much cleaner when it's done as a separate module.  Even
the check for forgetting to drop the lock on error paths would be better
pulled out into a separate module.

The new modular system will be cleaner, more precise and more flexible.

This code isn't ready for use yet, but I'm committing it because it
hopefully will be useful soon.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>