expressions: introduce expr_get_fake_or_real_parent_expr()
There are several places where we create fake assignments.
1) For complicated returns
2) For complicated parameters
3) For all function calls

The problem is that there are places where we want to say, "Is there
already an assignment?  And if so then treat it differently."  The problem is
they weren't detecting fake assignments.  Create a function which gets the
parent, whether or not it's fake.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>