flow/function_hooks: handle fake assignments better
The problem is that Smatch handles function assignments differently from
function calls which aren't assigned.  But then we started creating fake
function assignments because actually it's really handy to have
assignments instead of a lot of special cases.  So parameters are
converted.  frob(get()); becomes frob(fake = get());.  And then basically
all function calls which weren't a part of an assignment got a fake
assignment.

There were two problems with this:
1) We sometimes created two fake assignments so the function was parsed
   twice.
2) The code to detect fake assignments didn't handle some fake assignments
   correctly so if you had parentheses like frob((get())); that became
   frob(fake_param = (get())); and that wasn't detected as a function call
   so then that wasn't parsed at all.

With the new expr_get_fake_or_real_parent_expr() function then we can
handle this nicely.

Btw, one thing which is sort of weird is that += assignments are not
counted as a function assignment.  That doesn't cause any issues with the
current code but it doesn't seem right either.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>