conditions: fix "locked = !spin_trylock()"
The smatch_conditions.c module treats "foo = !bar;" as

    if (bar)
        foo = 0;
    else
        foo = 1;

As in it creates strees and all that.  This is especially nice where "bar"
is a function call.

The problem in this case was that if you set a state on one side of the
condition and not on the other, then the unset state should be &undefined.
But instead of that it was doing the merge based on the final stree so it
was a &merged state.  Save the pre_stree and do the merge based on that.

This commit also adds some comments and some calls to free_stree().

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>