locking: add ignored_modify to function_data
This doesn't cause a problem in real life, but if you wanted to print the
ignored lock in debug code then probably that would cause a problem.

I'm pretty proud of the fact that I noticed this issue before it caused a
crash on my system.  Normally, that didn't use to happen but I feel like I
have become smarter?  Probably I will look back on this day as a turning
point in my life.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>