function_hooks: don't double parse functions
So one thing that can happen is that we have a function call:

	function(&dest, frob());

And the function assigns dest = frob().  We should probably parse this
in the correct way.  But right now I'm just worried about double parsing
because I'm worried about "double lock" and "double free" bugs.  So
if we are in a fake parameter assignment then don't reparse the function.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>