locking_type: track when certain struct members are locked
One of the problems with tracking locking is that the variable names
change.  It could lock param->foo->bar->baz->lock and then unlock
baz->lock.  Also quite often it's hard to match names to parameters
especially when it's something like:

	p = find_lock(param, UID);
	lock(p->lock);

So I have the idea that instead of tracking the variable, we could just
track the struct member.  This might be useful to say that if we access
struct member p->foo then we need to hold lock p->lock.  All kinds of uses
really.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>