locking: silence certain mismatched locks
One thing that often happens is that we call the same lock different names.
param->foo->bar->baz->lock or baz->lock.  And the lock and the unlock
might not match.

I have the idea that the normal way this looks is that there is just one
lock which is confusing.  So let's silence the warning if we have two
states and one state is lock and the other is unlock.

But let's make it stricter than that.  Let's check smatch_locking_type.c
and check that it only has one state and for that state, it's back at the
start state.

This is very strict, and generally I like to write my checks to catch as
many bugs as possible.  Potentially in the future, I might ignore more
states.  But after one day of testing, it seems okay.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>