locking: replace check_locking.c with check_inconsistent_locking.c
I re-wrote the check_locking.c check to be modular.  This new check is a
lot simpler.  It produces only 30% of the warnings that the old code did
so it must be an improvement.

As I was committing this change, I discovered that check_call_tree.c needs
to print locks held.  I've commented that out for now.  I don't know of
anything which uses it.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>