assigned_expr: handle temporary macro variables better
I was debugging a false positive and I saw a function which did something
like:

struct foo *get_ptr(struct bar *p)
{
	return rcu_dereference_check(p->something);
}

And Smatch is supposed to record in the database that the return is
*p->something, but that wasn't happening.  And obviously I thought that
rcu_dereference_check() wasn't getting parsed correctly.  When you expand
the macro it expands to:

	fake_return = (({ type tmp = p->something; magic(); tmp }));

So I decided that recording that  fake_return is tmp is pointless because
the scope of tmp is tiny and Smatch discards it as soon as possible.

So I wrote this patch to solve the issue.  But when I went to test the code
some days later, the original bug had disappeared and everything was
working correctly.  (I rebuild my database every day and it disappeared
before I applied my fix).  I don't know what to think!  But the idea behind
this patch is sound and perhaps it helps with something so let's commit it.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>