locking: hack around guard() locks
The guard locks are tricky because of conditional locking.  The unlock
functions check if we are locked so it looks like maybe it's not a
unlock.

So far, what I've done is manually annotate the cleanup functions.  The
problem with this approach is that people keep adding new ones.

What this commit does is say that if there is a cleanup function then
don't record any transitions in the DB.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>