locking: create a more generic destructor function
People are creating lots of annoying destructor functions that all look
optional to a static checker so they have to be annotated manually.
Create an easier way to annotate them.

I should have made this use $ as a wild card instead of %s but obviously
it started as a hack that spread.  Will fix later.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>