flow: parse new scoped_guard() correctly
The re-work to scoped_guard() has landed in linux next and it creates tons
of "unreachable code" false positives, and means that a lot of functions
get marked as no_return functions.

The re-worked scoped_guard() macro does something like:

        for (struct lock __cleanup(destroy) = constructor;
             is_not_optional || took_lock();
             goto done_label)
                if (0) {
                        done_label: break;
                } else

Smatch has always ignored if (0) statements because it never occurred to me
that there would be a label statement in there.  But also because of the
way that Smatch handles loops, this was always going to need to be handled
specially.

I had done some work to parse the previous iterations of this re-work and
I accidentally committed the was_handle_backward_goto_in_scoped_guard()
part of that.  Delete this dead code.

Reported-by: Harshit Mogalapalli <harshit.m.mogalapalli@oracle.com>
Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>