summarize_errs.sh: fix skipping repeat warnings
The idea here was that some warning messages get printed a million times
which is an annoying waste of time.  So somehow, I thought I could use
`uniq -f 2` to filter out based on the repeated function name.  Which is
garbage for two reasons.  1)  Maybe it's two completely different warnings
in the same funciotn.  2)  That's not how uniq -f works...  To be honest,
I couldn't figure out how uniq -f works at all but it's not whatever I
expected.

Do the filtering more manually based on comparing strings.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>