db: fix a NULL dereference in get_param()
If you have a function where the parameter has no name:

void frob(int)  // <-- the parameter doesn't have a name but isn't used
{
	printf("hello\n");
}

This arg->ident is NULL.  There used to be one of these in reiserfs. In
Nov 2010 and I added a comment about it.  I was so puzzled that the
pointer was non-zero but still so tiny and I thought maybe there was
memory corruption going on.  But actually arg->ident->name is not a
pointer, it's an array.  14 years later, I'm finally able to remove my
"temporary hack".

Reported-by: Domingo Alvarez Duarte
Closes: https://github.com/error27/smatch/issues/13
Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>