divide_condition: delete this check
The divide condition check was basically because someone wrote a
completely wrong condition and I believe Smatch caught it.  But I was
like why are we writing "if (foo / 4) {" anyway?  That's the same thing
as "if (foo >= 4) {".  When I investigated, I found that a number of
people had intended to write "if (foo % 4) {" so the check seemed to be
useful.  But we've had it for 2 years and actually it's mostly not
useful.

It turns out that "if (foo / 8) {" can be idiomatic depending on the
subsystem.

Too many false positives.  Delete it.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>