param_cleared: don't mark certain structs as cleared
The main motivation for this change is that sometimes the struct isn't
defined so it always marks all the struct members as set because there
aren't any.  This breaks flow analysis because now variables are unknown
when actually they should be known.

But also we don't really need to use BUF_CLEAR for small structs.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>