locking: ignore schedule()
Ugh.  I give up.  I think where was an issue in schedule() and I fixed it.

But then Smatch kept marking schedule() as disabling IRQs.  Except I can't
figure out why and when I rebuild the database manually, it doesn't do
that.  But then when it gets rebuilt in the nightly build it keeps on going
back to the wrong behavior.  What?  Ok, I'm done with this and I'm just
going to hard code the function as a do-nothing function.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>