smatch: don't save smatch output to the .o files
I discovered this bug while trying to run Smatch on rsync.

Sparse now takes a -o option, but the problem is that the compiler
also wants to use the -o option.  And they both can't use it...
Smatch instead uses the --file-ouput option which creates foo.c.smatch
files.

Ignore the -o option.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>