db: improve handling of bool returns
There was a false positive in the kernel triggered by this code:

static inline bool queue_is_mq(struct request_queue *q)
{
        return q->mq_ops;
}
	...
	if (queue_is_mq(q)) {

What the code does is that it returns true/false depending on if
q->mq_ops is NULL or non-NULL.  We want Smatch to track the implications
correctly.

So for tracking the implications we need to break up the returns into
true/false instead of just one return.  The call_return_state_hooks()
has code to identify returning bool but it doesn't work correctly because
it was only looking at the type of the variable and not the type of the
function.

Then the other problem is that in smatch_function_hooks.c we create a
fake assignment to say that queue_is_mq() returns q->mq_ops.  But that's
not correct, it returns q->mq_ops casted to bool.  So we need to add the
cast.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>