no_null_check_on_mixed: warn about checking for IS_ERR() instead of NULL
What this does is that if we check a pointer for IS_ERR() but it can
be NULL and then later we dereference it without checking for NULL then
print a warning.

I just made this check stricter today because I changed to using the
add_dereference_hook() which makes it work across function boundaries.
The second changes is that originally I only printed a warning if the
pointer was "distinctly NULL" as in one ->possible was NULL.  Now, I
print a warning if a ->possible is "0,...".

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>