db/kernel.delete_return_states: delete incorrect return 1 states
The problem here is that these functions take a function pointer as
a parameter and return the result.  Obviously the callers knew what
function they are passing and what it returns.  I should filter by that
as well...  But instead of doing it properly, I'm going to hard code it.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>