no_null_check_on_mixed: ignore non-function assigns
One source of false positives is that Smatch doesn't understand how
threads work.  So you start with a file scope pointer and it's NULL
but then during probe it's either valid or an error pointer.  And the
users check for that.  But Smatch doesn't know that the probe() happened
so Smatch still thinks it could also be NULL.

Silence this by only warning about functions which return a mix of
error pointers and NULL.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>