unsigned_lt_zero: warn about unsigned_var <= 0
Checking for if (unsigned_var < 0) is a bit obvious because that's
an impossible thing.  Even this has been controversial recently.  :P

But checking for if (unsigned_var <= 0) is a even more controversial
because often the "fix" is just to change the <= to ==.  Which is not
a fix at all, but only a style change.

This change is very restricted and only warns when there is good reason
to think that the unsigned_var can be negative.  It only prints 3 warnings
across the entire kernel.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>