kernel.return_fixes: remove some NULL returns
These are functions which appear to return NULL but for practical
purposes they really don't.

nouveau_connector_create() triggers a WARN_ON() before the return NULL.
(Then the caller will crash.  Hooray!)

ubi_find_or_add_av() calls find_or_add_av() which can return NULL but
not with the flags that we pass to it.

iommu_attach_handle_get() masks away the bottom two bits and Smatch says
"well, it could be zero of > 3".  Which is true.  But actually it's a
valid pointer, hopefully.

The problem with cifs_sb_master_tlink() is that Smatch says it returns
unknown we pass it to other functions which check for NULL so Smatch says,
"Huh.  I guess it can be NULL".  Sometimes Smatch gets this correct and
says that the NULL is impossible but if it gets it wrong one time then
that leads to issues.  No, it can't be NULL.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>