bits: fix how if (x & BIT(0) is handled
It's easier to handle single bit tests.  And this is probably the
common case as well.  So let's say we have an if statement:

	if (x & BIT(0)) {

Then if it's possibly true, then that means BIT(0) need to set the ->set
bit on the true path and clean the ->possible bit on the false path.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>