inconsistent_locking: check bool type functions differently
For bool type functions we check that all the true returns are the same
and all the false returns are the same.  It's simpler than the default
check.  I think the motivation for this change was to be more accurate
instead of the motivation being to silence false positives?  I'm not
positive honestly.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>