return_efault: don't rely on the cross function DB
This check warns about if you return the result from copy_to_user()
instead of -EFAULT.

One false positive is from functions which never return negative error
codes.  My solution to this was to look up the error codes in the DB
and disable the warning if it always returned positive.  However the
problem with my code is that it disabled the warning if you didn't
have the cross function DB.  I hadn't intended to do that.

So to fix that we'll use the DB if it's there, otherwise we'll silence
the warning if the function is defined as unsigned.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>