buf_comparison: introduce get_kmalloc_pointer()
The smatch_buf_comparison.c modules looks at code like:

	p = kmalloc(sizeof(*p) * n, GFP_KERNEL);

And says that "n" is the number of elements in "p".  However, these days
the kmalloc function has been changed into a macro so it looks like:

p = ({ void *_res; _res = kmalloc(); _res});

And so we assign things to _res which is a void pointer instead of to
p and so Smatch says "I don't know the size of void pointers so I don't
know how many elements are in p".

The get_kmalloc_pointer() gives you p again.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>