param_limit: add a [h] for hard max limits
This adds a "[h]" to the end of PARAM_LIMIT for hard maxes.

I've been running with this code for a while.  It seems to kind of work.
I don't know that I'm super happy with it, but let's commit it to make the
diff smaller.

I really think I should pull the hard max stuff into a separate module.
Probably and estate should just be a range list...

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>