dereference: add devm_platform_ioremap_resource_byname()
You can't pass a NULL "name" pointer to
devm_platform_ioremap_resource_byname() or it will crash.  Technically,
there is a way that it might avoid crashing if there is an unrelated
error.  (This raises a good point because Smatch only warns about
functions which can't accept a NULL without crashing when really it should
be functions where the success path would crash.  #TODO).

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>