db/kernel.return_fixes: add __rb_reserve_next(), regmap_field_read() and __tag_set()
The __rb_reserve_next() function doesn't return NULL.
The regmap_field_read() function is recorded as returning non-zero, but
really it returns regular error codes.

The __tag_set() is interesting.  For some reason Smatch is recording it
as returning -1.  The -1 is because my config has the tag stuff as a no-op
so it's "addr | -1", but why is addr set to NULL?  Fix this by hard coding
it to unknown.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>