extra: fix a ->related bug in set_extra_nomod()
The problem here is that you have an assignment:

	a = b;

But then "b" changes while "a" stays the same.  So the ->related list for
"a" still says that "b" and "a" are the same. So when we discover something
about "a", like it's 0-10 for example, then we apply it to "b" as well.
Which is nonsense.  Checking that the related lists are the same is the
same as checking that the variables are the same.

This works, but I'm probably going to re-write this to use SSA soon
instead.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>