db/function_hooks: don't return "[== $0]"
Smatch, of course, evolved in a very ad hoc way.  So one of the things
that happen if is you have a functions like:

int one(int val)
{
	return val;
}

int two(struct my_struct *p)
{
	return p->a;
}

The return from the first function would be recorded as
"s32min-s32max[== $0]" and for the second it would be
"s32min-s32max[$0->a]"  The results were also parsed differently where
the first one would be handled when we had comparisons like:

	if (IS_ERR(p) == -PROBE_DEFER) {

And the second one would be handled when there were assignments and we
would create a fake assignment for it.

Change the first return so it becomes "s32min-s32max[$0]" and the result
is used for both comparisons and assignments.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>