db: re-order return state numbers
Imagine we do something like this:

void free(int **p)
{
	free(*p);
	*p = NULL;
}

In Smatch we don't record the order of things, we just have a list of
states at the end.  When the free happens, we say is this an unmodified
parameter?  Okay, we'll record that as freeing the parameter.  When we
assign *p = NULL that's tracked in param_set and we say "We assigned
the NULL to the pointer."  Those things recorded in the stree.

When we record them in the database they are sorted by type.  So
the PARAM_FREE has to come before the PARAM_SET.  Originally the order was
pretty random.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>