db: add all the data types that I have
How this normally works is that I have the idea that I'm going to
use a temporary type id so I pick a number in the 8000s, planning that I
will move it to the 1000s when it becomes permanent.  But it's like such
a big hassle to change the numbers that I just say "forget about it" and
leave it as 8000.

I think I'm going to just commit everything since there is no downside.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>