smdb.py: sort the results properly
The results from from the return_states are supposed to be sorted by
type.  (And return_id as well obviously).  The smdb.py script is
information only, the smatch_db.c code has been sorting this already
obviously or everything would be a lot more messed up than it is.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>