unused_ret: ignore fake assignments
What happens is that when we're parsing a function:

	foo = frob(bar);

Then it can result fake assignments like "foo = bar->baz;" and Smatch
says "Hey, why didn't we use the foo = bar() assignment?"  So we don't
want to record the fake assignments here, only the real ones.

I don't know if adding &assignment_list to the function data makes a
difference, but it's the correct thing to do.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>